generadorlinealcongruencial/pyproject.toml
[tool.poetry]
name = "pruba_generadorlinealcongruencial"
version = "0.1.0"
description = ""
authors = ["qleoz12 <kintaroloop@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.1"
Jinja2 = "^3.0.3"
matplotlib = "^3.5.1"
pycairo = "^1.20.1"
PyQt5 = "^5.15.6"



[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
